Move op stack chain specs to new crate reth-optimism-chainspec
