+++ crates/chainspec/Cargo.toml
@@ -27,7 +27,6 @@ alloy-trie.workspace = true
 # op
 op-alloy-rpc-types = { workspace = true, optional = true }
 
-
 # misc
 auto_impl.workspace = true
 once_cell.workspace = true
@@ -56,5 +55,6 @@ std = []
 arbitrary = [
     "alloy-chains/arbitrary"
 ]
+test-utils = []
