+++ crates/node/core/Cargo.toml
@@ -38,6 +38,7 @@ reth-network-peers.workspace = true
 reth-consensus-common.workspace = true
 reth-prune-types.workspace = true
 reth-stages-types.workspace = true
+reth-optimism-chainspec = { workspace = true, features = ["optimism"], optional = true }
 
 # ethereum
 alloy-genesis.workspace = true
@@ -87,6 +88,7 @@ optimism = [
     "reth-provider/optimism",
     "reth-rpc-types-compat/optimism",
     "reth-rpc-eth-api/optimism",
+    "dep:reth-optimism-chainspec",
 ]
 # Features for vergen to generate correct env vars
 jemalloc = []
