+++ crates/optimism/chainspec/src/constants.rs
@@ -1,0 +1,11 @@
+//! OP stack variation of chain spec constants.
+
+//------------------------------- BASE MAINNET -------------------------------//
+
+/// Max gas limit on Base: <https://basescan.org/block/17208876>
+pub const BASE_MAINNET_MAX_GAS_LIMIT: u64 = 105_000_000;
+
+//------------------------------- BASE SEPOLIA -------------------------------//
+
+/// Max gas limit on Base Sepolia: <https://sepolia.basescan.org/block/12506483>
+pub const BASE_SEPOLIA_MAX_GAS_LIMIT: u64 = 45_000_000;
