+++ crates/optimism/chainspec/src/op.rs
@@ -1,0 +1,40 @@
+//! Chain specification for the Optimism Mainnet network.
+
+#[cfg(not(feature = "std"))]
+use alloc::sync::Arc;
+#[cfg(feature = "std")]
+use std::sync::Arc;
+
+use alloy_chains::Chain;
+use alloy_primitives::{b256, U256};
+use once_cell::sync::Lazy;
+use reth_chainspec::{BaseFeeParams, BaseFeeParamsKind, ChainSpec};
+use reth_ethereum_forks::{EthereumHardfork, OptimismHardfork};
+use reth_primitives_traits::constants::ETHEREUM_BLOCK_GAS_LIMIT;
+
+/// The Optimism Mainnet spec
+pub static OP_MAINNET: Lazy<Arc<ChainSpec>> = Lazy::new(|| {
+    ChainSpec {
+        chain: Chain::optimism_mainnet(),
+        // genesis contains empty alloc field because state at first bedrock block is imported
+        // manually from trusted source
+        genesis: serde_json::from_str(include_str!("../res/genesis/optimism.json"))
+            .expect("Can't deserialize Optimism Mainnet genesis json"),
+        genesis_hash: Some(b256!(
+            "7ca38a1916c42007829c55e69d3e9a73265554b586a499015373241b8a3fa48b"
+        )),
+        paris_block_and_final_difficulty: Some((0, U256::from(0))),
+        hardforks: OptimismHardfork::op_mainnet(),
+        base_fee_params: BaseFeeParamsKind::Variable(
+            vec![
+                (EthereumHardfork::London.boxed(), BaseFeeParams::optimism()),
+                (OptimismHardfork::Canyon.boxed(), BaseFeeParams::optimism_canyon()),
+            ]
+            .into(),
+        ),
+        max_gas_limit: ETHEREUM_BLOCK_GAS_LIMIT,
+        prune_delete_limit: 10000,
+        ..Default::default()
+    }
+    .into()
+});
