+++ crates/optimism/chainspec/src/op_sepolia.rs
@@ -1,0 +1,38 @@
+//! Chain specification for the Optimism Sepolia testnet network.
+
+#[cfg(not(feature = "std"))]
+use alloc::sync::Arc;
+#[cfg(feature = "std")]
+use std::sync::Arc;
+
+use alloy_chains::{Chain, NamedChain};
+use alloy_primitives::{b256, U256};
+use once_cell::sync::Lazy;
+use reth_chainspec::{BaseFeeParams, BaseFeeParamsKind, ChainSpec};
+use reth_ethereum_forks::{EthereumHardfork, OptimismHardfork};
+use reth_primitives_traits::constants::ETHEREUM_BLOCK_GAS_LIMIT;
+
+/// The OP Sepolia spec
+pub static OP_SEPOLIA: Lazy<Arc<ChainSpec>> = Lazy::new(|| {
+    ChainSpec {
+        chain: Chain::from_named(NamedChain::OptimismSepolia),
+        genesis: serde_json::from_str(include_str!("../res/genesis/sepolia_op.json"))
+            .expect("Can't deserialize OP Sepolia genesis json"),
+        genesis_hash: Some(b256!(
+            "102de6ffb001480cc9b8b548fd05c34cd4f46ae4aa91759393db90ea0409887d"
+        )),
+        paris_block_and_final_difficulty: Some((0, U256::from(0))),
+        hardforks: OptimismHardfork::op_sepolia(),
+        base_fee_params: BaseFeeParamsKind::Variable(
+            vec![
+                (EthereumHardfork::London.boxed(), BaseFeeParams::optimism_sepolia()),
+                (OptimismHardfork::Canyon.boxed(), BaseFeeParams::optimism_sepolia_canyon()),
+            ]
+            .into(),
+        ),
+        max_gas_limit: ETHEREUM_BLOCK_GAS_LIMIT,
+        prune_delete_limit: 10000,
+        ..Default::default()
+    }
+    .into()
+});
