+++ crates/optimism/node/Cargo.toml
@@ -37,6 +37,7 @@ reth-rpc-eth-types.workspace = true
 reth-rpc-eth-api.workspace = true
 reth-optimism-rpc.workspace = true
 reth-rpc.workspace = true
+reth-optimism-chainspec = { workspace = true, features = ["optimism"] }
 
 # async
 async-trait.workspace = true
