+++ crates/optimism/node/tests/e2e/utils.rs
@@ -2,11 +2,12 @@ use std::sync::Arc;
 
 use alloy_genesis::Genesis;
 use reth::{rpc::types::engine::PayloadAttributes, tasks::TaskManager};
-use reth_chainspec::{ChainSpecBuilder, BASE_MAINNET};
+use reth_chainspec::ChainSpecBuilder;
 use reth_e2e_test_utils::{transaction::TransactionTestContext, wallet::Wallet, NodeHelperType};
 use reth_node_optimism::{
     node::OptimismAddOns, OptimismBuiltPayload, OptimismNode, OptimismPayloadBuilderAttributes,
 };
+use reth_optimism_chainspec::BASE_MAINNET;
 use reth_payload_builder::EthPayloadBuilderAttributes;
 use reth_primitives::{Address, B256};
 use tokio::sync::Mutex;
