+++ crates/primitives/Cargo.toml
@@ -20,6 +20,7 @@ reth-trie-common.workspace = true
 revm-primitives = { workspace = true, features = ["serde"] }
 reth-chainspec = { workspace = true, optional = true }
 reth-codecs = { workspace = true, optional = true }
+reth-optimism-chainspec = { workspace = true, optional = true, features = ["optimism"] }
 
 # ethereum
 alloy-primitives = { workspace = true, features = ["rand", "rlp"] }
@@ -99,6 +100,7 @@ optimism = [
     "reth-chainspec/optimism",
     "reth-ethereum-forks/optimism",
     "revm-primitives/optimism",
+    "dep:reth-optimism-chainspec",
 ]
 alloy-compat = ["reth-primitives-traits/alloy-compat", "dep:alloy-rpc-types"]
 test-utils = ["reth-primitives-traits/test-utils"]
